Insertion Functionality Complete
User may add a new object or attempt to pull exact data from another object. If the object in question is in the database, the database will create a new object based on most of the same data from the selected OSN object and a few unique values from the current object. If the object is not valid, the user will be asked if they would like to create a new object. Is yes, the object creation menu is populated; else, the Main Menu is displayed again.